add bilinear interpolation
Interpolate elevation between points in the
data store. This provides smoother elevation
points in slippy map tiles, which dont exactly
align with elevation points in the data store.

+++ src/main/kotlin/com/greensopinion/elevation/processor/elevation/BlockMapper.kt
@@ -18,7 +18,7 @@ class BlockMapper(
             y = floor((90.0 - coordinates.latitude) / blockSize.degrees).toInt() - yOffset
         )
         val min = Coordinates(
-            latitude = -((block.y + yOffset) * blockSize.degrees - 90.0) - blockSize.degrees,
+            latitude = -((block.y + yOffset.toDouble()) * blockSize.degrees - 90.0) - blockSize.degrees,
             longitude = (block.x - xOffset.toDouble()) * blockSize.degrees - 180.0
         )
         val max = Coordinates(
